# internal/chat/domain/model/room.go
package model

type RoomModel struct {
	ID           string                `json:"id" example:"70383d59-3b82-4ebb-bdc5-adf7a036dda5"`
	Name         string                `json:"name" example:"General Chat"`
}